compiler: aliases.Candidate -> ir.ExprGeometry
