+++ devito/ir/support/__init__.py
@@ -1,5 +1,5 @@
-from .utils import *  # noqa
 from .vector import *  # noqa
+from .utils import *  # noqa
 from .basic import *  # noqa
 from .space import *  # noqa
 from .guards import *  # noqa
